Refactor composer and chat TSX hotspots
